pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sanskrit-env"
version = "1.0.0"
description = "Sanskrit Manuscript Interpretation Environment — an RL environment for resolving structural ambiguity in Sanskrit manuscripts."
requires-python = ">=3.11"
license = {text = "Apache-2.0"}
dependencies = [
    "openenv-core>=0.2.2",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "huggingface-hub>=0.23.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.0.0",
]

[project.optional-dependencies]
dev = []

[project.scripts]
server = "server.app:main"
